requirements.txt

# 3rd-party packages:
python-dotenv
openai>=1.0.0
PyPDF2>=3.0.0
gradio[oauth,mcp]==5.33.1
pydantic>=2.0
requests